requirements.txt
torch
easyocr
torchvision
supervision==0.18.0
openai==1.3.5
transformers
ultralytics==8.3.61
azure-identity
numpy
opencv-python
opencv-python-headless
gradio
dill
accelerate
timm
einops==0.8.0
paddlepaddle
paddleocr
ipykernel
Pillow
nvidia-pyindex
nvidia-cuda-runtime-cu12
python-dotenv==1.0.1